split AVSs to separate package
