+++ README.md
@@ -149,7 +149,7 @@ For an example see https://github.com/etherfi-protocol/etherfi-avs-operator-CLI/
 The CLI command should be a simple wrapper that forwards data to the package you implement
 in the following step
 
-### 3. Implement core logic in a new package `src/{my_avs}`
+### 3. Implement core logic in a new package `src/avs/{my_avs}`
 Please also place any abi's and generated bindings in this package.
 For an example see https://github.com/etherfi-protocol/etherfi-avs-operator-CLI/blob/witness-chain/src/witnesschain/witnesschain.go
