+++ bin/avs-cli/eigenda/eigenda.go
@@ -9,7 +9,7 @@ import (
 	registryCoordinator "github.com/Layr-Labs/eigenda/contracts/bindings/RegistryCoordinator"
 
 	"github.com/dsrvlabs/etherfi-avs-operator-tool/bindings"
-	"github.com/dsrvlabs/etherfi-avs-operator-tool/src/eigenda"
+	"github.com/dsrvlabs/etherfi-avs-operator-tool/src/avs/eigenda"
 	"github.com/dsrvlabs/etherfi-avs-operator-tool/src/etherfi"
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/urfave/cli/v3"
