+++ bin/avs-cli/witness-chain/register_watchtower.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/dsrvlabs/etherfi-avs-operator-tool/src/witnesschain"
+	"github.com/dsrvlabs/etherfi-avs-operator-tool/src/avs/witnesschain"
 	"github.com/urfave/cli/v3"
 )
