+++ src/avs/eoracle/eoracle.go
