+++ src/avs/eoracle/eoracle_service_manager.abi
