+++ src/avs/witnesschain/witness_chain_operator_registry.abi
